Add all hibernate entities for database

+++ src/main/resources/db/changelog/db.changelog-1.0.sql
@@ -82,7 +82,7 @@ CREATE TABLE IF NOT EXISTS review
     hotel_id    INT REFERENCES hotel (id) ON DELETE CASCADE NOT NULL,
     user_id     INT REFERENCES users (id)                   NOT NULL,
     created_at  TIMESTAMP                                   NOT NULL,
-    rating      INT                                         NOT NULL,
+    rating      SMALLINT                                    NOT NULL,
     description TEXT
 );
 --rollback DROP TABLE review;
